+++ src/components/modules/toolbar/blockSettings.ts
@@ -219,7 +219,7 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
       tool.toolbox.forEach((toolboxItem) => {
         result.push({
           icon: toolboxItem.icon,
-          title: toolboxItem.title,
+          title: I18n.t(I18nInternalNS.toolNames, toolboxItem.title),
           name: tool.name,
           closeOnActivate: true,
           onActivate: async () => {
@@ -240,6 +240,7 @@ export default class BlockSettings extends Module<BlockSettingsNodes> {
     if (convertToItems.length > 0) {
       items.push({
         icon: IconReplace,
+        name: 'convert-to',
         title: I18n.ui(I18nInternalNS.ui.popover, 'Convert to'),
         children: {
           searchable: true,