+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `Fix` – The onChange callback won't be fired when editor is initialized in the Read-Only mode
 - `Fix` – Convert To supports i18n again
+- `Fix` – Prevent form submit on inline tool click
 
 ### 2.30.1